+++ bindings/python/examples/how_tos/nfts/burn_nft.py
@@ -22,5 +22,4 @@ nftId = "0xf95f4d5344217a2ba19a6c19a47f97d267edf8c4d76a7b8c08072ad35acbebbe"
 
 # Send transaction.
 transaction = account.prepare_burn_nft(nftId).send()
-print(f'Block sent: {os.environ["EXPLORER_URL"]}/block/{transaction["blockId"]}')
-
+print(f'Block sent: {os.environ["EXPLORER_URL"]}/block/{transaction.blockId}')